+++ Changes.md
@@ -2,6 +2,7 @@
 
 * Visualizer: Key bindings to alter Quality (Ctrl+Q) and Hue (Ctrl+H)
 * Remote: Automatically change Hue over time
+* Remote: Set preset auto change time
 
 ## v3.4 (2025-11-22)